lifting useEffect() hook up
